fix(misc): parser validation, portability, overflow, and SAH normalization
Add Catch2 test suite (64 tests) and GitHub Actions CI.

Bugs fixed:
1. Parser accepted missing 'rendermode', invalid mode strings,
   missing camera object, missing width/height/fov, and
   non-positive dimensions. Added validation with descriptive
   error messages for each case.
2. Unknown shape type in JSON leaked the just-allocated Material.
   Added else branch that deletes mat and pops from materials_
   before continuing.
3. Material::load_texture didn't check stream state after each
   >> extraction — truncated/corrupt PPM silently filled
   remaining pixels with 0. Added stream check with warning,
   reset to no-texture state on failure. Also added per-pixel
   clamping to [0, max_val].
4. ImageWriter::validateParameters computed width*height as int
   before casting to size_t — overflows for large images
   (e.g. 50000×50000). Changed to cast each dimension
   individually before multiplying.
5. Node::sah_splits SAH cost was unnormalized (raw area×count).
   Divided by parent bbox surface_area() so cost is comparable
   across axis splits.
6. M_PI replaced with project constant pi in Material.h,
   Cylinder.h, and Sphere.h for portability (M_PI is non-standard
   POSIX extension).
7. main.cpp missing #include <iomanip> — failed to compile on
   some toolchains.

Tests added:
- tests/brdf_tests.cpp: Lambert/FancyBRDF sampling and evaluation
- tests/vector_tests.cpp: Vector3 arithmetic, dot, cross, norm
- tests/radiance_tests.cpp: tone mapping (linear, reinhard, filmic)
- tests/bbox_tests.cpp: AABB intersection, union, surface_area
- tests/object_tests.cpp: Sphere/Triangle intersection, normals
- tests/integration_tests.cpp: full scene render with ground-truth
  comparison

CI: .github/workflows/test.yml — build + test on push/PR.
CMakeLists.txt: added Catch2 test target and sanitizer build flags.

+++ src/ImageWriter.cpp
@@ -11,7 +11,7 @@ bool ImageWriter::validateParameters() const {
         std::cerr << "Error: Width and height must be positive." << std::endl;
         return false;
     }
-    if (pixelData.size() != static_cast<size_t>(width * height)) {
+    if (pixelData.size() != static_cast<size_t>(width) * static_cast<size_t>(height)) {
         std::cerr << "Error: Pixel data size does not match width and height." << std::endl;
         return false;
     }
